-- fixed windows warnings

+++ usrsctplib/user_recv_thread.c
@@ -1136,7 +1136,7 @@ recv_thread_init(void)
 #endif
 #if defined(INET)
 	if (SCTP_BASE_VAR(userspace_rawsctp) == -1) {
-		if ((SCTP_BASE_VAR(userspace_rawsctp) = socket(AF_INET, SOCK_RAW, IPPROTO_SCTP)) < 0) {
+		if ((SCTP_BASE_VAR(userspace_rawsctp) = (int) socket(AF_INET, SOCK_RAW, IPPROTO_SCTP)) < 0) {
 #if defined(__Userspace_os_Windows)
 			SCTPDBG(SCTP_DEBUG_USR, "Can't create raw socket for IPv4 (errno = %d).\n", WSAGetLastError());
 #else
@@ -1187,7 +1187,7 @@ recv_thread_init(void)
 		}
 	}
 	if (SCTP_BASE_VAR(userspace_udpsctp) == -1) {
-		if ((SCTP_BASE_VAR(userspace_udpsctp) = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP)) < 0) {
+		if ((SCTP_BASE_VAR(userspace_udpsctp) = (int) socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP)) < 0) {
 #if defined(__Userspace_os_Windows)
 			SCTPDBG(SCTP_DEBUG_USR, "Can't create socket for SCTP/UDP/IPv4 (errno = %d).\n", WSAGetLastError());
 #else
@@ -1251,7 +1251,7 @@ recv_thread_init(void)
 #endif
 #if defined(INET6)
 	if (SCTP_BASE_VAR(userspace_rawsctp6) == -1) {
-		if ((SCTP_BASE_VAR(userspace_rawsctp6) = socket(AF_INET6, SOCK_RAW, IPPROTO_SCTP)) < 0) {
+		if ((SCTP_BASE_VAR(userspace_rawsctp6) = (int) socket(AF_INET6, SOCK_RAW, IPPROTO_SCTP)) < 0) {
 #if defined(__Userspace_os_Windows)
 			SCTPDBG(SCTP_DEBUG_USR, "Can't create socket for SCTP/IPv6 (errno = %d).\n", WSAGetLastError());
 #else
@@ -1324,7 +1324,7 @@ recv_thread_init(void)
 		}
 	}
 	if (SCTP_BASE_VAR(userspace_udpsctp6) == -1) {
-		if ((SCTP_BASE_VAR(userspace_udpsctp6) = socket(AF_INET6, SOCK_DGRAM, IPPROTO_UDP)) < 0) {
+		if ((SCTP_BASE_VAR(userspace_udpsctp6) = (int) socket(AF_INET6, SOCK_DGRAM, IPPROTO_UDP)) < 0) {
 #if defined(__Userspace_os_Windows)
 			SCTPDBG(SCTP_DEBUG_USR, "Can't create socket for SCTP/UDP/IPv6 (errno = %d).\n", WSAGetLastError());
 #else
